refactor drafts and devices
