+++ src/utils/api/index.js
@@ -1,6 +1,7 @@
 import Account from '@/utils/api/account.js';
 import Blacklist from '@/utils/api/blacklist.js';
 import Conversations from '@/utils/api/conversations.js';
+import Devices from '@/utils/api/devices.js';
 import Drafts from '@/utils/api/drafts.js';
 import Folders from '@/utils/api/folders.js';
 import Messages from '@/utils/api/messages.js';
@@ -11,6 +12,7 @@ export {
     Account,
     Blacklist,
     Conversations,
+    Devices,
     Drafts,
     Folders,
     Messages,
